server/backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Cors==5.0.0
Flask-Limiter==3.8.0
Flask-Talisman==1.1.0
Flask-Migrate==4.1.0
Flask-Swagger-UI==4.11.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
stripe==11.4.1
SQLAlchemy==2.0.37
Werkzeug==3.0.6
supabase-py==0.0.2 